Plain text conversion is done after markers substitution #54769
We used to convert first and then substitute markers, but that workflow
allowed re-introducing HTML tags via markers, or the conversion could
break the strict marker syntax.

Now the conversion to plain text is the very last step before sending,
simplifying logic (only one substituion in HTML version) and avoiding all
pitfalls mentionned above. The only downside is we had to drop the
template based converter since it did not fit in the new workflow (or it
would have had no marker substitution at all). Anyway it is believed to be
insignificant enough to be dropped.