Substitute marker in URL before redirection of spied links
Target URL do not exists in spied link anymore. Instead the
non-substituted URL is stored in DB. And when we need to redirect
to target URL it is retrived from DB and markers are substituted.

That way statistics are coherent for a single non-subsistued URL
(instead of 1 link and 1 click per user, we have 1 link for everybody
with many clicks). Final redirection is correct because substitution is
done right before the redirect. And finally the redirection system is
not open to exploit since the only URL possible to be redirect to are
stored in DB.

FIX #79