Support for multi-domains
We used to hardcode the the PID 0 to create a fake TSFE to build
frontend URIs from the backend. This used to work for single domain
install, but totally crash the BE module (and more) for multi-domains
install.

We now don't hardcode anything and instead use the respective object's
PID to generate link. The result is that Newsletter is able to generate
links across domains correctly.