Expose "update" button in Extension Manager
This partially revert the automated update system contributed by
@marvin-martian because it was not triggered when updating the
extension via composer. Since composer is the way to go, we must
provide a way to apply updates for that workflow. Unfortunately
composer installation never use any kind of hooks/signals so it cannot
be automated. So the last resort is to use the legacy way of showing
an "update" button in Extension Manager. However we improved the
situation by only showing it when necessary and doing the update
immediately on clicking (without any kind of confirmation).