Use ugly URL for XML/CSV export to avoid breaking backend
We gave up pretty URL for XML/CSV export of recipient list in the TCA,
because it very frequently broke the entire backend for many users and
was very unreliable and hard to maintain properly.

So instead of relying on TYPO3 core functionalities for URL building,
we semi-hardcoded a simpler URL builder that should only be used from
TCA context. This is much simpler and does not really matter to have
pretty URL in backend anyway, especially for URL that are bound to be
temporary (because of easily changing `authCode`).

The good news is that it brings back export for TYPO3 < 7.6 users.

Fixes #136