Don't test on TYPO3 8.6
TYPO3 8.6 introduced a huge breaking change in testing component
by extracting it from core package and changing its namespace. It
seems unlikely that we will be able to support both LTS (7 and 8)
at the same time with the same testing code.

So for now we disabled testing for newer version than 8.6 and we will
re-evaluate the situation once the 8 LTS is actually released.

Also see https://forge.typo3.org/issues/79025#note-18