Default language is `0` and not `-1` for recipient list
As [documented here](https://docs.typo3.org/typo3cms/FrontendLocalizationGuide/TyposcriptConfiguration/Index.html#config-sys-language-uid),
the proper value for default language is `0` and not an empty string or `-1`. This
used to break newsletter sending if the recipient list had `-1` selected
(by mistake or intentionnaly). Then when fetching content TYPO3 correctly raised
an exception stating that the language was invalid.