Loosen check for parallel runs
We looked for an email sent in the last 15 seconds to assume a parallel
spool was running. This was not enough for very edge cases. So we now
look for an email that has started being sent in the last 30 seconds.
This should give way enough margins even for edge cases to prevent parallel
runs of spools.

Fixes #147