Further speed up for email sending
By ignoring linkAuthCode and now also emailAuthCode we are able to
speed up even more the email sending process. That means that it is
no longer possible to include linkAuthCode or emailAuthCode in pretty
URL (and cache them). But this never make much senses to begin with
as those two parameters are specifically meant to make the URL unique,
and thus caching them is counter-productive.

With Newsletter 3.2.0 sending an email with 88 links to
100 recipients in a single run generated ~33'000 SQL queries. A vast
majority of those were made by RealURL. With this commit, under the
same conditions we are at ~1'000 SQL queries. That is still a lot, but
way less than the gigantic previous count.

#146