Remove standard fields (delete, hidden, etc.) on email and newsletter
Those standard fields were never supported at all by our code. And it should
have never existed in the first place, because both record types are created
semi-automatically and are usually not meant for manual edition and especially
never for manual soft-deletion. It does not make sense to soft-delete an email,
because it will never be able to be soft-undeleted properly.

Fixes #170