Remove typo3/cms-lang dependency
While we still rely on that package we cannot declare it as dependency,
because it does not exist in TYPO3 9.5 anymore. Declaring like we did would
force to install an outdated extension for nothing.

Old TYPO3 versions will get the package installed via core. New TYPO3
versions won't have the package at all.